// server/src/models/User.ts
export interface User {
    ID: number;
    user_login: string;
    user_nicename: string;
    user_email: string;
    user_url: string;
    user_registered: string;
    user_activation_key: string;
    user_status: number;
    display_name: string;
    user_role: string;
    enrolment_date: string;
}